refacto: split main
